Update FMT to fix two issues (installer packaging + installer extraction)
Add some weird retry policy on the download method when a weird ssl handshake error happened for some dark and unknown reasons

Update OptiFine integration: no more length check because it's not longer provided.

+++ src/main/java/fr/flowarg/flowupdater/download/DownloadList.java
@@ -58,7 +58,7 @@ public class DownloadList
 
         if (this.optiFine != null)
         {
-            this.downloadInfo.totalToDownloadBytes.set(this.downloadInfo.totalToDownloadBytes.get() + (long)(this.optiFine.getSize()));
+            this.downloadInfo.totalToDownloadBytes.set(this.downloadInfo.totalToDownloadBytes.get() + this.optiFine.getSize());
             this.downloadInfo.totalToDownloadFiles.incrementAndGet();
         }
         this.init = true;
